create game_team_league inherited class
