+++ test/test_helper.rb
@@ -11,3 +11,4 @@ require './lib/stat_tracker'
 require './lib/team_collection'
 require './lib/team'
 require './lib/game_team_game'
+require './lib/game_team_league'
